# PERF: improve plotting performance by not stringifying all x data