--- a/asv_bench/benchmarks/plotting.py
+++ b/asv_bench/benchmarks/plotting.py
@@ -10,21 +10,46 @@ except ImportError:
     from pandas.tools.plotting import andrews_curves
 
 
+class Plotting(object):
+    goal_time = 0.2
+
+    def setup(self):
+        import matplotlib
+        matplotlib.use('Agg')
+        self.s = Series(np.random.randn(1000000))
+        self.df = DataFrame({'col': self.s})
+
+    def time_series_plot(self):
+        self.s.plot()
+
+    def time_frame_plot(self):
+        self.df.plot()
+
+
 class TimeseriesPlotting(object):
     goal_time = 0.2
 
     def setup(self):
         import matplotlib
         matplotlib.use('Agg')
-        self.N = 2000
-        self.M = 5
-        self.df = DataFrame(np.random.randn(self.N, self.M), index=date_range('1/1/1975', periods=self.N))
+        N = 2000
+        M = 5
+        idx = date_range('1/1/1975', periods=N)
+        self.df = DataFrame(np.random.randn(N, M), index=idx)
+
+        idx_irregular = pd.DatetimeIndex(np.concatenate((idx.values[0:10],
+                                                         idx.values[12:])))
+        self.df2 = DataFrame(np.random.randn(len(idx_irregular), M),
+                             index=idx_irregular)
 
     def time_plot_regular(self):
         self.df.plot()
 
     def time_plot_regular_compat(self):
         self.df.plot(x_compat=True)
+
+    def time_plot_irregular(self):
+        self.df2.plot()
 
 
 class Misc(object):