Closes #18236

Currently when plotting all x / index data are converted to strings, while you typically only need a few tick labels. So when you have a lot of data, this cause the pandas plotter to be hugely slower than needed (and than a pure matplotlib one)

On master:

```
[ 33.33%] ··· Running plotting.Plotting.time_frame_plot                        4.43s
[ 50.00%] ··· Running plotting.Plotting.time_series_plot                       4.19s
[ 66.67%] ··· Running plotting.TimeseriesPlotting.time_plot_irregular          72.2±0.2ms
[ 83.33%] ··· Running plotting.TimeseriesPlotting.time_plot_regular            108±0.5ms
[100.00%] ··· Running plotting.TimeseriesPlotting.time_plot_regular_compat     71.5±0.8ms
```

with this branch:

```
[ 33.33%] ··· Running plotting.Plotting.time_frame_plot                        132±20ms
[ 50.00%] ··· Running plotting.Plotting.time_series_plot                       71.4±30ms
[ 66.67%] ··· Running plotting.TimeseriesPlotting.time_plot_irregular          58.7±0.5ms
[ 83.33%] ··· Running plotting.TimeseriesPlotting.time_plot_regular            96.9±2ms
[100.00%] ··· Running plotting.TimeseriesPlotting.time_plot_regular_compat     57.4±0.8ms
```

So for very simple plot when from 4s to ca 100ms (which is much closer to the pure matplotlib performance)