@jreback 2017-11-20
lgtm. needs a whatsnew, prob ok for 0.21.1