@TomAugspurger 2017-11-20
https://github.com/pandas-dev/pandas/issues/18236