@jorisvandenbossche 2017-11-20
Ah, yes, didn't see that one, but from the description this PR is exactly fixing that :-)