@jorisvandenbossche 2017-11-20
@TomAugspurger any comments on the content of the PR?